refactor source files
